docs: some tsdocing
