+++ packages/services/src/Logger/LoggerChannelRegistry.ts
@@ -3,12 +3,23 @@ import { createLogger } from 'winston';
 
 import { TransportFactory } from './TransportFactory';
 
-import type { ChannelConfig, LoggerConfiguration, LoggerLevel, TransportConfig, WinstonInstance } from './types';
+import type { ChannelConfig, LoggerConfiguration, LoggerLevel, TransportConfig, WinstonInstance } from './aTypes';
 
-const defaultLevel: LoggerLevel = Envapter.isDevelopment ? 'silly' : Envapter.isStaging ? 'debug' : 'info';
-
+/**
+ * Manages Winston logger instances per channel with caching.
+ *
+ * Handles channel configuration, transport creation,
+ * and environment-aware defaults.
+ * @internal
+ */
 export class LoggerChannelRegistry {
     private static _instance: LoggerChannelRegistry;
+
+    private readonly DEFAULT_LEVEL: LoggerLevel = Envapter.isDevelopment
+        ? 'silly'
+        : Envapter.isStaging
+          ? 'debug'
+          : 'info';
 
     private config: LoggerConfiguration = {
         defaultChannel: 'default',
@@ -26,6 +37,9 @@ export class LoggerChannelRegistry {
         this.transportFactory = new TransportFactory();
     }
 
+    /**
+     * Gets the singleton instance of the registry.
+     */
     public static get instance(): LoggerChannelRegistry {
         // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
         return (this._instance ??= new LoggerChannelRegistry());
@@ -34,9 +48,9 @@ export class LoggerChannelRegistry {
     private getDefaultChannelConfig(name: string): ChannelConfig {
         return {
             name,
-            level: defaultLevel,
+            level: this.DEFAULT_LEVEL,
             transports: [
-                { type: 'console', level: defaultLevel, format: 'pretty' },
+                { type: 'console', level: this.DEFAULT_LEVEL, format: 'pretty' },
                 {
                     type: 'file',
                     level: Envapter.isDevelopment ? 'debug' : 'info',
@@ -68,15 +82,29 @@ export class LoggerChannelRegistry {
         };
     }
 
+    /**
+     * Updates global logger configuration and clears cache.
+     *
+     * @param config - Partial configuration to merge with existing settings
+     */
     public configure(config: Partial<LoggerConfiguration>): void {
         this.config = { ...this.config, ...config, channels: { ...this.config.channels, ...(config.channels ?? {}) } };
         this.cache.clear();
     }
 
+    /**
+     * Returns the name of the default channel.
+     */
     public getDefaultChannel(): string {
         return this.config.defaultChannel;
     }
 
+    /**
+     * Gets or creates a Winston logger instance for the given channel.
+     *
+     * @param channel - Channel name to get logger for
+     * @returns Configured Winston logger instance
+     */
     public get(channel: string): WinstonInstance {
         const cached = this.cache.get(channel);
         if (cached) return cached;
@@ -86,7 +114,7 @@ export class LoggerChannelRegistry {
             this.config.channels[channel]
         );
 
-        const effectiveLevel = channelConfig.level ?? defaultLevel;
+        const effectiveLevel = channelConfig.level ?? this.DEFAULT_LEVEL;
         const transports = (channelConfig.transports ?? []).map((transportConfig: TransportConfig) =>
             this.transportFactory.build({
                 channel,
